@@ -21,7 +21,7 @@
   ],
   "require": {
     "ext-json": "*",
-    "php": "^7.0|^8.0",
+    "php": "^8.0",
     "phpoffice/phpspreadsheet": "^1.18",
     "illuminate/support": "5.8.*|^6.0|^7.0|^8.0|^9.0",
     "box/spout": "^3.3",
