@@ -3,6 +3,7 @@
 namespace Maatwebsite\Excel;
 
 use Box\Spout\Reader\Common\Creator\ReaderEntityFactory;
+use Box\Spout\Writer\WriterInterface;
 use Box\Spout\Writer\Common\Creator\WriterEntityFactory;
 use Maatwebsite\Excel\Cache\CacheManager;
 use Maatwebsite\Excel\Concerns\WithCustomValueBinder;
@@ -18,18 +19,12 @@ use Maatwebsite\Excel\Factories\WriterFactory;
 use Maatwebsite\Excel\Files\RemoteTemporaryFile;
 use Maatwebsite\Excel\Files\TemporaryFile;
 use Maatwebsite\Excel\Files\TemporaryFileFactory;
-use PhpOffice\PhpSpreadsheet\Cell\Cell;
-use PhpOffice\PhpSpreadsheet\IOFactory;
-use PhpOffice\PhpSpreadsheet\Spreadsheet;
 
 class SpoutWriter implements Writer
 {
     use DelegatedMacroable, HasEventBus;
 
-    /**
-     * @var Spreadsheet
-     */
-    protected $spreadsheet;
+    protected array $spreadsheet;
 
     /**
      * @var object
@@ -52,67 +47,28 @@ class SpoutWriter implements Writer
     }
 
     /**
-     * @param object $export
-     * @param string $writerType
-     *
-     * @return TemporaryFile
-     * @throws \PhpOffice\PhpSpreadsheet\Exception
-     */
-    public function export($export, string $writerType): TemporaryFile
-    {
-        $this->open($export);
-
-        $sheetExports = [$export];
-        if ($export instanceof WithMultipleSheets) {
-            $sheetExports = $export->sheets();
-        }
-
-        foreach ($sheetExports as $sheetExport) {
-            $this->addNewSheet()->export($sheetExport);
-        }
-
-        return $this->write($export, $this->temporaryFileFactory->makeLocal(null, strtolower($writerType)), $writerType);
-    }
-
-    /**
-     * @param object $export
-     *
-     * @return $this
-     */
-    public function open($export)
-    {
-        $this->exportable = $export;
-
-        if ($export instanceof WithEvents) {
-            $this->registerListeners($export->registerEvents());
-        }
-
-        $this->exportable  = $export;
-        $this->spreadsheet = new Spreadsheet;
-        $this->spreadsheet->disconnectWorksheets();
-
-        if ($export instanceof WithCustomValueBinder) {
-            Cell::setValueBinder($export);
-        }
-
-        $this->handleDocumentProperties($export);
-
-        $this->raise(new BeforeExport($this, $this->exportable));
-
-        return $this;
-    }
-
-    /**
      * @param TemporaryFile $tempFile
      * @param string        $writerType
      *
      * @return Writer
-     * @throws \PhpOffice\PhpSpreadsheet\Reader\Exception
      */
-    public function reopen(TemporaryFile $tempFile, string $writerType)
+    public function reopen(TemporaryFile $tempFile, string $writerType): SpoutWriter
     {
-        $reader = ReaderEntityFactory::createReaderFromFile($tempFile->sync()->getLocalPath());
-        $this->spreadsheet = $reader;
+        $path = $tempFile->sync()->getLocalPath();
+        $reader = ReaderEntityFactory::createReaderFromFile($path);
+
+        $ss = [];
+        // read everything into memory 
+        foreach ($reader->getSheetIterator() as $sheetIndex => $sheet) {
+            $ss[$sheetIndex] = [];
+            
+            foreach ($sheet->getRowIterator() as $row) {
+                array_push($ss[$sheetIndex], $row);
+            }
+        }
+        $reader->close();
+
+        $this->spreadsheet = $ss;
 
         return $this;
     }
@@ -123,47 +79,40 @@ class SpoutWriter implements Writer
      * @param string        $writerType
      *
      * @return TemporaryFile
-     * @throws \PhpOffice\PhpSpreadsheet\Writer\Exception
-     * @throws \PhpOffice\PhpSpreadsheet\Exception
      */
     public function write($export, TemporaryFile $temporaryFile, string $writerType): TemporaryFile
     {
         $this->exportable = $export;
-
-        $this->spreadsheet->setActiveSheetIndex(0);
-
+        
         $this->raise(new BeforeWriting($this, $this->exportable));
+        \Log::info("write using spout");
+        
+        $path = $temporaryFile->sync()->getLocalPath();
+        
+        $writer = WriterEntityFactory::createWriterFromFile($path);
+        $writer->openToFile($path);
 
-        $writer = WriterFactory::make(
-            $writerType,
-            $this->spreadsheet,
-            $export
-        );
+        foreach ($this->spreadsheet as $sheetIndex => $sheet) {
+            if ($sheetIndex !== 1) {
+                $writer->addNewSheetAndMakeItCurrent();
+            }
 
-        $writer->save(
-            $path = $temporaryFile->getLocalPath()
-        );
-
+            foreach ($sheet->getRowIterator() as $row) {
+                // ... and copy each row into the new spreadsheet
+                $writer->addRow($row);
+            }
+        }
+        $writer->close();
+        
         if ($temporaryFile instanceof RemoteTemporaryFile) {
             $temporaryFile->updateRemote();
+            $temporaryFile->deleteLocalCopy();
         }
 
-        $this->spreadsheet->disconnectWorksheets();
+        $this->clearListeners();
         unset($this->spreadsheet);
-        app(CacheManager::class)->flush();
 
         return $temporaryFile;
-    }
-
-    /**
-     * @param int|null $sheetIndex
-     *
-     * @return Sheet
-     * @throws \PhpOffice\PhpSpreadsheet\Exception
-     */
-    public function addNewSheet(int $sheetIndex = null)
-    {
-        return new Sheet($this->spreadsheet->createSheet($sheetIndex));
     }
 
     /**
@@ -171,89 +120,16 @@ class SpoutWriter implements Writer
      */
     public function getDelegate()
     {
-        return $this->spreadsheet;
-    }
-
-    /**
-     * @return $this
-     */
-    public function setDefaultValueBinder()
-    {
-        Cell::setValueBinder(
-            app(config('excel.value_binder.default', DefaultValueBinder::class))
-        );
-
-        return $this;
+        return $this->spreadsheet ;
     }
 
     /**
      * @param int $sheetIndex
      *
      * @return Sheet
-     * @throws \PhpOffice\PhpSpreadsheet\Exception
      */
     public function getSheetByIndex(int $sheetIndex): Sheet
     {
-        return new Sheet($this->getDelegate()->getSheet($sheetIndex));
-    }
-
-    /**
-     * @param string $concern
-     *
-     * @return bool
-     */
-    public function hasConcern($concern): bool
-    {
-        return $this->exportable instanceof $concern;
-    }
-
-    /**
-     * @param object $export
-     */
-    protected function handleDocumentProperties($export)
-    {
-        $properties = config('excel.exports.properties', []);
-
-        if ($export instanceof WithProperties) {
-            $properties = array_merge($properties, $export->properties());
-        }
-
-        if ($export instanceof WithTitle) {
-            $properties = array_merge($properties, ['title' => $export->title()]);
-        }
-
-        $props = $this->spreadsheet->getProperties();
-
-        foreach (array_filter($properties) as $property => $value) {
-            switch ($property) {
-                case 'title':
-                    $props->setTitle($value);
-                    break;
-                case 'description':
-                    $props->setDescription($value);
-                    break;
-                case 'creator':
-                    $props->setCreator($value);
-                    break;
-                case 'lastModifiedBy':
-                    $props->setLastModifiedBy($value);
-                    break;
-                case 'subject':
-                    $props->setSubject($value);
-                    break;
-                case 'keywords':
-                    $props->setKeywords($value);
-                    break;
-                case 'category':
-                    $props->setCategory($value);
-                    break;
-                case 'manager':
-                    $props->setManager($value);
-                    break;
-                case 'company':
-                    $props->setCompany($value);
-                    break;
-            }
-        }
+        return new SpoutSheet($this->spreadsheet[$sheetIndex]);
     }
 }
