@@ -2,7 +2,7 @@
 
 namespace Maatwebsite\Excel\Events;
 
-use Maatwebsite\Excel\Sheet;
+use Maatwebsite\Excel\PhpSpreadsheetSheet;
 
 class AfterSheet extends Event
 {
@@ -20,7 +20,7 @@ class AfterSheet extends Event
      * @param  Sheet  $sheet
      * @param  object  $exportable
      */
-    public function __construct(Sheet $sheet, $exportable)
+    public function __construct(PhpSpreadsheetSheet $sheet, $exportable)
     {
         $this->sheet      = $sheet;
         $this->exportable = $exportable;
@@ -29,7 +29,7 @@ class AfterSheet extends Event
     /**
      * @return Sheet
      */
-    public function getSheet(): Sheet
+    public function getSheet(): PhpSpreadsheetSheet
     {
         return $this->sheet;
     }
